refactor(runner): Move grading helper into support package
Place extracted runner grading helpers under a dedicated runner_support package for subsequent runner refactor modules.
